// src/app/models/utils/preview-attachment.enum.ts
export enum PreviewAttachmentEnum {
  PREVIEW_PDF = 'https://cdn.chattigo.com/assets/img/attachment/preview/PDF.svg',
  PREVIEW_TINY_PDF = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/PDF.svg',
  PREVIEW_WORD = 'https://cdn.chattigo.com/assets/img/attachment/preview/DOCX.svg',
  PREVIEW_TINY_WORD = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/DOCX.svg',
  PREVIEW_EXCEL = 'https://cdn.chattigo.com/assets/img/attachment/preview/XLS.svg',
  PREVIEW_TINY_EXCEL = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/XLS.svg',
  PREVIEW_PPT = 'https://cdn.chattigo.com/assets/img/attachment/preview/PPT.svg',
  PREVIEW_TINY_PPT = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/PPT.svg',
  PREVIEW_AUDIO = 'https://cdn.chattigo.com/assets/img/attachment/preview/Audio.svg',
  PREVIEW_TINY_AUDIO = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/Audio.svg',
  PREVIEW_ONE = 'https://cdn.chattigo.com/assets/img/attachment/preview/DOCX.svg',
  PREVIEW_TINY_ONE = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/DOCX.svg',
  PREVIEW_TEXT = 'https://cdn.chattigo.com/assets/img/attachment/preview/TXT.svg',
  PREVIEW_TINY_TEXT = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/TXT.svg',
  PREVIEW_OTHER = 'https://cdn.chattigo.com/assets/img/attachment/preview/Default.svg',
  PREVIEW_TINY_OTHER = 'https://cdn.chattigo.com/assets/img/attachment/preview-tiny/Default.svg',
}